Rename util.ts→owner-document.ts
